Added Tag export, Added unit tests, Refactored export logic (skipping briefing data, parse filename for mission data)
