--- tS_MissionInfoGrabber/src/ru/ts/missionlist/grabber/entity/BriefingData.java
@@ -1,13 +1,0 @@
-package ru.ts.missionlist.grabber.entity;
-
-public class BriefingData {
-    private final String text;
-
-    public BriefingData(String text) {
-        this.text = text;
-    }
-
-    public String getText() {
-        return text;
-    }
-}
