+++ tS_MissionInfoGrabber/src/test/resources/CO16-Wastelands-1B.ProvingGrounds_PMC/dzn_tSFramework/Modules/Briefing/tSF_briefing.sqf
@@ -1,0 +1,51 @@
+//     tSF Briefing
+// Do not modify this part
+#define BRIEFING		_briefing = [];
+#define TOPIC(NAME) 	_briefing pushBack ["Diary", [ NAME,
+#define END			]];
+#define ADD_TOPICS	for "_i" from (count _briefing) to 0 step -1 do {player createDiaryRecord (_briefing select _i);};
+//
+//
+// Briefing goes here
+
+BRIEFING
+
+TOPIC("I. Обстановка:")
+"14 марта 2012 года. Республика Такистан.<br /><br />Активность инсургентов резко возросла за последние месяцы, после открытых заявлений правительства США о введении санкций против Такистана.<br /><br />В связи с этим армия возобновляет патрулирования удаленных районов страны, чтобы пресечь попытки поставок оружия и наемников из пограничных областей."
+END
+
+TOPIC("А. Враждебные силы:")
+"Маловероятно - инсургенты (племена роялистов), численностью до взвода."
+END
+
+TOPIC("Б. Дружественные силы:")
+"2х моторизированные отделения Армии Такистана"
+END
+
+TOPIC("II. Задание:")
+"- Пройти по контрольным точкам КТ 1, КТ 2, КТ 3.<br />- В случае обнаружения инсургентов - сообщить штабу, при возможности - уничтожить."
+END
+
+TOPIC("III. Выполнение:")
+"Порядок контрольных точек не имеет значения.<br />Это патрульная операция, поэтому в первую очередь мы должны обнаружить присутствие противника, а не вступать с ним в бой. Отступление - приемлемое решение.<br />"
+END
+
+TOPIC("IV. Поддержка:")
+"2х Land Rover (M2HB)<br />2x Land Rover"
+END
+
+TOPIC("V. Сигналы:")
+"PL NET - LR 50<br />1'1 - SR CH 1<br />1'2 - SR CH 2"
+END
+
+TOPIC("VI. Замечания:")
+"- Нет"
+END
+
+if ((serverCommandAvailable '#logout') || !(isMultiplayer) || isServer) then {
+TOPIC("VII. GSO notes:")
+"- Поддержка недоступна ни игрокам, ни противникам (никакие AC-130 не прилетят)!"
+END
+};
+
+ADD_TOPICS
