+++ tS_MissionInfoGrabber/src/test/resources/co21-succesive-patrol-1c.Kunduz/dzn_tSFramework/Modules/Briefing/tSF_briefing.sqf
@@ -1,0 +1,17 @@
+//     tSF Briefing
+// Do not modify this part
+#define BRIEFING		_briefing = [];
+#define TOPIC(NAME) 	_briefing pushBack ["Diary", [ NAME,
+#define END			]];
+#define ADD_TOPICS	for "_i" from (count _briefing) to 0 step -1 do {player createDiaryRecord (_briefing select _i);};
+//
+//
+
+TAGS([])
+
+// Briefing goes here
+
+BRIEFING
+
+
+ADD_TOPICS
